Deploy/app_usedCars/GCP_Airflow/requirements.txt
psycopg2-binary==2.9.9
apache-airflow==2.7.3
apache-airflow-providers-google==10.20.0
apache-airflow-providers-github==2.6.2
apache-airflow[google_auth]
pyarrow
unidecode==1.3.7
numpy==1.23.5
pandas==1.5.3
mlflow==2.12.2
requests==2.31.0
python-multipart==0.0.9
uvicorn==0.27.1
xgboost==2.0.3
catboost==1.0.6
lightgbm==4.3.0
scikit-learn==1.4.2
scipy==1.13.0
matplotlib==3.7.0
seaborn==0.12.2
dill==0.3.8
future==1.0.0
shap==0.42.0
urllib3==1.26.18
streamlit==1.31.1
path==16.14.0
